requirements.txt
azure-ai-documentintelligence==1.0.0b1
azure-ai-formrecognizer
azure-cognitiveservices-speech
azure-common==1.1.28
azure-core==1.30.0
azure-identity==1.15.0
azure-search-documents==11.4.0
bs4==0.0.2
langchain==0.1.4
langchain_community==0.0.16
langchain_core==0.1.17
langchain_openai
markdown==3.3.4
matplotlib
opencv-python==4.9.0.80
openai==1.12.0
pyperclip
pypdfium2==4.27.0
python-docx==0.8.11
python-dotenv==1.0.1
python-pptx==0.6.21
PyMuPDF>=1.18.14
PyPDF2==1.26.0
requests==2.28.2
streamlit==1.31.0
streamlit-webrtc==0.47.1
unstructured
youtube-transcript-api==0.6.1
plotly

